fix for empty refsubject in csv and shift to local afni
After update @LISA, there's no longer a working afni module. Requested a new install, but for now you should download the binaries to your home folder and add them to your path in .bashrc. This is required for the preprocessing workflow (and any other that uses afni, but I don't there are others in standard NHP-BIDS right now)